requirements.txt.txt
tweepy==3.10.0
requests==2.26.0
python-dotenv==0.19.1
Pillow==8.4.0